Made game playable for 5 rounds
Created a function called "game" which runs the function "playRound" 5 times using a for-loop. "game" function also stores the amount of times the player/ computer wins to display whether the computer or player won overall across 5 games.

Added comments in the code which accurately describes the role of each function or subsection of code.

Switched from "console.log()" to "alert()" statements to make the game more playable.