Game ends at 5 wins
Created a function to update player scores, instead of doing it separately on each click event for the buttons.

At the end of the function, program checks to see if either player or computer has reached 5 wins, in which case the scores are reset and the appropriate winner is announced via result textview.